% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{as.data.frame.Interval}
\alias{as.data.frame.Interval}
\title{Convert an Interval object to a data frame}
\usage{
\method{as.data.frame}{Interval}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An \code{Interval} object.}

\item{row.names}{Passed to \code{data.frame()}.
Either NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{Passed to \code{data.frame()}.
A single logical value, if \code{TRUE}, setting row names and converting column names (to syntactic names: see make.names) is optional.
\code{!optional} is passed to \code{check.names} in \code{data.frame()}.
Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{data.frame()}.}
}
\value{
A data frame with two columns: \code{start} and \code{end}.
}
\description{
Split an \code{Interval} object into a data frame with two columns: \code{start} and \code{end}.
}
