% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{as_interval}
\alias{as_interval}
\title{Create an Interval object from a date range}
\usage{
as_interval(date_range = NULL, start = NULL, end = NULL)
}
\arguments{
\item{date_range}{(Optional - required if \code{start} and \code{end} are not provided).
A vector of 2 dates, datetimes, or characters
(formatted as "YYYY-MM-DD HH:MM:SS" UTC time or "now" for the current time) representing the start and end time period.
NA values will be replaced with 1970-01-01 or "now" for the first and second values, respectively.}

\item{start}{(Optional - required if \code{date_range} is not provided).
A date, datetime, or character (formatted as "YYYY-MM-DD HH:MM:SS" UTC time or "now" for the current time) representing the start of the date range.}

\item{end}{(Optional - required if \code{date_range} is not provided).
A date, datetime, or character (formatted as "YYYY-MM-DD HH:MM:SS" UTC time or "now" for the current time) representing the end of the date range.}
}
\value{
An \code{Interval} object representing the date range.
}
\description{
\code{Interval} objects represent a range of time and  can be created using \verb{\%--\%} from the \code{lubridate} package.
\code{\link[=as_interval]{as_interval()}} makes this 'pipe friendly', allowing you to create an \code{Interval} object from a vector of 2 dates, datetimes, or characters
(formatted as "YYYY-MM-DD HH:MM:SS" UTC time or "now" for the current time).
}
