% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual__buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Generating a list for construction a network using the function forceNetwork from the networkD3 package}
\usage{
buildNetwork(df, w)
}
\description{
Generating a list for construction a network using the function forceNetwork from the networkD3 package
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
