% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_communities}
\alias{canada_communities}
\title{Canadian communities from OpenStreetMap}
\format{
\subsection{\code{canada_communities}}{

An \code{sf} data frame with 11,313 rows and 5 columns:
\describe{
\item{osm_id}{OpenStreetMap ID}
\item{name}{Name of community}
\item{prov_terr}{Province or territory of community}
\item{type}{Type of community (city, town, village, hamlet)}
\item{geometry}{\code{sf} geometry column for community point location}
...
}
}
}
\source{
\url{https://openstreetmap.org}
}
\usage{
canada_communities
}
\description{
Locations and names of all city, town, village, and hamlet communities in Canada from OpenStreetMap as of October 2025.
}
\keyword{datasets}
