% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url_exists.R
\name{check_urls_exist}
\alias{check_urls_exist}
\title{Check if URLs exist and can be reached}
\usage{
check_urls_exist(urls, quiet = FALSE, ...)
}
\arguments{
\item{urls}{A character vector of URLs to check.}

\item{quiet}{A logical value indicating if warnings should be suppressed when not able to resolve host names for a URL. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[httr:HEAD]{httr::HEAD()}}.}
}
\value{
A logical vector indicating if the corresponding URL exists.
}
\description{
Performs a HEAD request on each URL and returns \code{TRUE} if the status code is 200, \code{FALSE} otherwise.
}
