% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Create a connection to a database}
\usage{
connect_to_database(name, type = NULL, path = NULL, ...)
}
\arguments{
\item{name}{The name of (and optionally path to) the database to create.
If the file extension does not match the type, it will be added.}

\item{type}{The type of database to create.
If NULL (the default), the type will be derived from \code{name}, defaulting to "sqlite".}

\item{path}{The location to save the database file to.
If NULL (the default), the path will be derived from \code{name}, defaulting to the current working directory.}

\item{...}{Additional arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\description{
Currently supports sqlite, duckdb, and PostgreSQL databases.
Must provide connection details for PostgreSQL databases (see \code{\link[RPostgres:Postgres]{RPostgres::Postgres()}}).

Be sure to call \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} on the database connection when finished working with it.
}
\examples{
\dontrun{
  "example.sqlite" |> create_database(type = "sqlite")
  db <- connect_to_database(name = "example.sqlite")
  DBI::dbDisconnect(db)
}
}
