% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between common units}
\usage{
convert_units(x, from, to)
}
\arguments{
\item{x}{A numeric vector to convert}

\item{from}{A string indicating the current unit of \code{x}. Examples: "cm", "mm", "in", "ft", "m", "km", "mi", "yd", "nm", "mi", "mile", "km", "mi", "yd", "ft", "in", "mm", "cm"}

\item{to}{A string indicating the desired unit of \code{x}. Examples: "cm", "mm", "in", "ft", "m", "km", "mi", "yd", "nm", "mi", "mile", "km", "mi", "yd", "ft", "in", "mm", "cm"}
}
\value{
A numeric vector of \code{x} converted to the desired unit
}
\description{
A wrapper around \code{\link[units:units]{units::set_units()}} to convert between common units
Run \code{\link[units:valid_udunits]{units::valid_udunits()}} for a list of valid units
}
\examples{
convert_units(c(1, 2, 3), from = "cm", to = "km")
convert_units(c(1, 2, 3), from = "degF", to = "degC")
convert_units(c(1, 2, 3), from = "ft", to = "m")
convert_units(c(1, 2, 3), from = "oz", to = "L")
convert_units(c(1, 2, 3), from = "m/s", to = "km/h")
convert_units(c(1, 2, 3), from = "ug/m3", to = "g/km3")
}
