% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinf__countPanGenes.R
\name{countPanGenes}
\alias{countPanGenes}
\title{Counting genes in a pangenome when randomly sampling a series of genomes}
\usage{
countPanGenes(P, core = 0.95, hard_core = 0.99, permutations = 20)
}
\arguments{
\item{P}{A binary presence-absence matrix with the dimension of n(genes) x n(isolates)}

\item{core}{Threshold defining hard/soft-core genes (Default: >=0.95)}

\item{hard_core}{Threshold defining hard-core genes (Default: >=0.99)}

\item{permutations}{Number of sampling for each sample size (Default: 20; minumum: 3)}
}
\value{
A list of two lists ("num" and "avg"): one for gene counts and the other for mean gene counts.
}
\description{
This function is useful for creating gene-accumulation curves, also known as rarefaction curves.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
