% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_transaction.R
\name{db_transaction}
\alias{db_transaction}
\title{Submit a transaction to a database, rolling back if an error occurs}
\usage{
db_transaction(db, ...)
}
\arguments{
\item{db}{A database connection or path to database (if sqlite or duckdb file extension).}

\item{...}{A code block to run in the transaction. The full request needs to be done here, including connecting to \code{db}.}
}
\value{
The result of the code block
}
\description{
Submit a transaction to a database, rolling back if an error occurs
}
