% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_database_entries.R
\name{delete_database_entries}
\alias{delete_database_entries}
\title{Delete specific entries from a database table.}
\usage{
delete_database_entries(db, table_name, entry_keys)
}
\arguments{
\item{db}{The database connection or path to the database file (if sqlite or duckdb file extension).}

\item{table_name}{The name of the table to delete entries from.}

\item{entry_keys}{A data frame or named vector with names matching those in the table and entries equal to the those in the rows to delete.}
}
\description{
Delete specific entries from a database table.
}
