% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file_lines.R
\name{delete_file_lines}
\alias{delete_file_lines}
\title{Delete select lines from a file inline}
\usage{
delete_file_lines(file_path, line_range)
}
\arguments{
\item{file_path}{The path to the file to modify.}

\item{line_range}{A numeric vector of length 2 indicating the start and end of the range of lines to delete. Line numbers start at 1, which in most cases is the header line - ensure you don't delete the header by accident.}
}
\description{
Uses sed to delete select lines from a file inline -
faster that loading the file into memory,
then deleting the lines, and then writing the file back out.
}
