% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__df2matrix.R
\name{df2matrix}
\alias{df2matrix}
\title{Coverts a dataframe that corresponds to either the upper/lower triangle in a matrix into a matrix.}
\usage{
df2matrix(df, diag = 1, dimension = NULL, symmetric = FALSE,
  replace.na = NA)
}
\description{
Variables in the data frame: c(name1, name2, value)
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
