% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo__distsPerClade.R
\name{distsPerClade}
\alias{distsPerClade}
\title{Summarising phylogenetic distances for each clade in a rooted bifurcating phylogenetic tree}
\usage{
distsPerClade(tr, m = NULL, dec = 8)
}
\arguments{
\item{tr}{An input bifurcating phylogenetic tree}

\item{m}{A distance matrix with row and column names matching to tip labels of the input tree}

\item{dec}{Number of decimals kept for distance summaries}
}
\value{
A data frame summarising phylogenetic distances for each clade
}
\description{
This function extracts clades from a rooted bifurcating phylogenetic tree and reports
the minimum, median, and the maximum of phylogenetic distances of each clade. Phylogenetic distances
may be SNP counts (neighbour-joining tree), number of substitutions per site (maximum-likelihood tree),
and so forth. The input tree will be midpoint rooted if it is unrooted.
}
