% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if_enough.R
\name{do_if_enough}
\alias{do_if_enough}
\title{Apply a function if enough values are provided}
\usage{
do_if_enough(x, FUN, ..., .min_length = 0)
}
\arguments{
\item{x}{A vector of values}

\item{FUN}{A function to be applied to x}

\item{...}{Additional arguments to be passed to FUN}

\item{.min_length}{The minimum length of \code{x} to be considered enough}
}
\value{
The output of FUN applied to x if enough values are provided, NA otherwise
}
\description{
\code{do_if_enough} provides a simple way to apply a function if enough values are provided.
}
\examples{
do_if_enough(c(1, 2, 3), mean, min_length = 2)
do_if_enough(c(1, 2, NA), mean, min_length = 2)
do_if_enough(c(1, NA, NA), mean, min_length = 2)
}
