% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise_mean.R
\name{elementwise_mean}
\alias{elementwise_mean}
\title{Vectorized element-wise mean, optionally ignoring NA values}
\usage{
elementwise_mean(..., digits = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{Vectors to compute the element-wise mean of. Must all be the same length.}

\item{digits}{Number of digits to round to (optional, default is \code{NULL} for no rounding).}

\item{na.rm}{If TRUE, NA values are ignored.}
}
\value{
Vector of the element-wise mean of \code{...} with the same length as vectors provided.
}
\description{
Vectorized element-wise mean, optionally ignoring NA values
}
