% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_epiLinksPatientIsolates.R
\name{epiLinksPatientIsolates}
\alias{epiLinksPatientIsolates}
\title{Identifying epidemiological links with isolates from a specific patient}
\usage{
epiLinksPatientIsolates(p, c, L, D, G, d_max = 90, e_max = 90)
}
\arguments{
\item{p}{A patient ID that can be found in L}

\item{c}{Genetic cluster of interest}

\item{L}{A data frame of four columns: "Isolate", "Cluster", "Patient", "Lab"; "Cluster" stands for genetic clusters, and "Lab" may represent a hospital or healthcare trust.}

\item{D}{A symmetric matrix of maximum days between possible collection dates of isolates}

\item{G}{A symmetric matrix of genetic distances}

\item{d_max}{maximum days to define an epidemiological link}

\item{e_max}{maximum days to define an episode of infection or carriage}
}
\description{
Finding out epidemiologically linked isolates within the same each genetic cluster c as isolates of a specific patient.
}
