% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sf_coords.R
\name{extract_sf_coords}
\alias{extract_sf_coords}
\title{Add sf coordinates and crs as columns}
\usage{
extract_sf_coords(sf_obj, keep_sf = TRUE)
}
\arguments{
\item{sf_obj}{An object of class \code{sf}, typically created by \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{keep_sf}{A logical value indicating if the return object should remain an \code{sf} object or be converted to a \code{data.frame}}
}
\value{
If \code{keep_sf} is:
\itemize{
\item \code{TRUE} \code{sf_obj} with two additional numeric columns (\code{x} and \code{y}) and a character column (\code{crs}).
\item \code{FALSE}: same as above, but converted to a \code{data.frame}
}
}
\description{
\code{extract_sf_coords} adds the x and y coordinates and the crs as additional columns to an \code{sf} object, with the option to convert it to a \code{data.frame}.
}
\examples{
cities <- data.frame(
  name = c("Nanaimo", "Port Moody", "Prince George"),
  x = c(-124.0531, -122.8519, -122.7949),
  y = c(49.1633, 49.2844, 53.8934)
)
cities_sf <- cities |>
  sf::st_as_sf(coords = c("x", "y"), crs = "WGS84")
cities_sf |>
  extract_sf_coords()
}
