% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_unzip.R
\name{get_and_unzip}
\alias{get_and_unzip}
\title{Get a zip file and unzip it}
\usage{
get_and_unzip(
  zip_url,
  local_path = file.path(".", basename(zip_url)),
  unzip_dir = dirname(local_path),
  quiet = FALSE,
  cache = TRUE,
  ...
)
}
\arguments{
\item{zip_url}{URL of zip file to download}

\item{local_path}{Where to save the zip file}

\item{unzip_dir}{Directory to unzip the contents to}

\item{quiet}{suppress messages?}

\item{cache}{Skip downloading the zip if it already exists?}

\item{...}{Additional arguments passed to \code{utils::download.file()}}
}
\value{
The path(s) to the unzipped file(s)
}
\description{
Get a zip file and unzip it
}
