% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_communities.R
\name{get_communities}
\alias{get_communities}
\title{Get a communities within a region from OpenStreetMap}
\usage{
get_communities(
  bbox,
  types = c("city", "town", "village", "hamlet"),
  timeout = 6000,
  quiet = TRUE
)
}
\arguments{
\item{bbox}{Usually the name of an area (ie. "Fort St. John, BC").
See \code{\link[osmdata:getbb]{osmdata::getbb()}} for more information.}

\item{types}{A character vector of types of communities to search for. Default is c("city", "town", "village", "hamlet"). See \href{https://wiki.openstreetmap.org/wiki/Key:place}{https://wiki.openstreetmap.org/wiki/Key:place}.}

\item{timeout}{The time in seconds to wait for a response from the OSM API. Default is 6000 seconds (10 minutes).}

\item{quiet}{A logical value indicating whether non-critical warnings/messages should be suppressed. Default is TRUE.}
}
\value{
An \code{sf} data frame containing the communities in the bounding box
with columns "osm_id", "type" (community type), "name", "population"/"population_source" (if available), and "geometry".
}
\description{
Get a communities within a region from OpenStreetMap
}
\examples{
get_communities("Fort St. John, BC")

get_communities("Yukon, Canada")
}
