% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_index.R
\name{get_file_index}
\alias{get_file_index}
\title{Get a file index from a URL}
\usage{
get_file_index(url, size_unit = "kB", timeout = 2)
}
\arguments{
\item{url}{The URL to fetch the file index from}

\item{size_unit}{The unit to express the file sizes in. Defaults to "kB".}

\item{timeout}{The timeout in seconds for the httr request. Defaults to 2.}
}
\description{
This function fetches a file index from a URL for a remote directory.
The file index is a data frame with columns for the file/directory
name, last modified date, size, and type (file or dir).
}
\examples{
get_file_index("https://aqmap.ca/aqmap/outputs/")
}
