% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interval.R
\name{get_interval}
\alias{get_interval}
\title{Determine distance(s) between values}
\usage{
get_interval(x, most_common = TRUE, na.rm = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{A numeric, Date, POSIXt, or POSIXct vector.}

\item{most_common}{A logical value indicating whether the most common interval (default) should be returned or all intervals and their frequencies in a sorted data frame.}

\item{na.rm}{A logical value indicating whether NAs should be removed before calculating the interval frequencies.
Default is \code{FALSE}.}

\item{quiet}{A logical value indicating whether warnings should be suppressed and output should be invisible.
Default is \code{FALSE}.}
}
\value{
\itemize{
\item If \code{most_common = TRUE} (default), a single numeric or difftime value.
\item If \code{most_common = FALSE}, a sorted data frame with columns \code{interval} (numeric or difftime) and \code{frequency} (integer).
}
}
\description{
\code{get_interval} finds the most common interval in a vector. If there are multiple intervals of the same frequency, it takes the first one in the sorted vector and a warning is raised.
}
\examples{
get_interval(c(1, 3, 5:10))
get_interval(as.Date(c("2020-01-01", "2020-01-03", "2020-01-05", "2020-01-06")))
get_interval(as.POSIXct(c("2020-01-01 00:00:00", "2020-01-01 00:00:02", "2020-01-01 00:00:04", "2020-01-01 00:00:05")))
get_interval(c(0, 3, 5, 7:10), most_common = FALSE)
}
