% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precision.R
\name{get_precision}
\alias{get_precision}
\title{Get the number of decimal places in a numeric vector}
\usage{
get_precision(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the same length as \code{x}, giving the number of decimal places for each element in \code{x}.
}
\description{
Get the number of decimal places in a numeric vector
}
\examples{
get_precision(1:5)
get_precision(c(1.2, 1.23, 1.234))
}
