% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season}
\alias{get_season}
\title{Get the season from a date}
\usage{
get_season(
  dates = Sys.time(),
  as_factor = FALSE,
  include_year = FALSE,
  include_months = FALSE,
  use_autumn = FALSE
)
}
\arguments{
\item{dates}{A Date or POSIXct vector of dates to get the season for.
Defaults to \code{Sys.time()}.}

\item{as_factor}{A logical indicating whether the output should be a factor or a character vector.
Defaults to \code{FALSE}.}

\item{include_year}{A logical indicating whether the output should include the year of the season or not.
Defaults to \code{FALSE}.}

\item{include_months}{A logical indicating whether the output should include the months of the season or not.
Defaults to \code{FALSE}.}

\item{use_autumn}{A logical indicating whether "Autumn" should be used instead of "Fall".
Defaults to \code{FALSE}.}
}
\value{
If \code{as_factor = FALSE} (default), a character vector with the season(s) and optionally the year/months.
If \code{as_factor = TRUE}, a factor of the above with chronological levels.
}
\description{
\code{get_season} takes a date (or set of dates) and returns the corresponding season(s).
}
