% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_step.R
\name{get_step}
\alias{get_step}
\title{Determine most common distance between values}
\usage{
get_step(x, most_common = TRUE, na.rm = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{A numeric, Date, POSIXt, or POSIXct vector.}

\item{most_common}{A logical value indicating whether the most common interval (default) should be returned or all intervals and their frequencies in a sorted data frame.}

\item{na.rm}{A logical value indicating whether NAs should be removed before calculating the interval frequencies.
Default is \code{FALSE}.}

\item{quiet}{A logical value indicating whether warnings should be suppressed and output should be invisible.
Default is \code{FALSE}.}
}
\value{
\itemize{
\item If \code{most_common = TRUE} (default), a single numeric or difftime value.
\item If \code{most_common = FALSE}, a sorted data frame with columns \code{interval} (numeric or difftime) and \code{frequency} (integer).
}
}
\description{
Finds the most common interval/step in a numeric-like vector.
Especially useful for time series data that has occasional time gaps.
If there are multiple intervals of the same frequency and \code{most_common = TRUE}, it takes the first one in the sorted vector and a warning is raised.
}
\examples{
get_step(c(1, 3, 5:10))
get_step(as.Date(c("2020-01-01", "2020-01-03", "2020-01-05", "2020-01-06")))
get_step(as.POSIXct(c("2020-01-01 00:00:00", "2020-01-01 00:00:02", "2020-01-01 00:00:04")))
get_step(c(0, 3, 5, 7:10), most_common = FALSE)
}
