% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc__getBinaryCoefficient.R
\name{getBinaryCoefficient}
\alias{getBinaryCoefficient}
\title{Calculate association or similarity coefficient for two binary variables
and test for their independence.}
\usage{
getBinaryCoefficient(
  x,
  y,
  type = "Jaccard",
  hypo.test = "chisq",
  alt = "two.sided"
)
}
\description{
A Jaccard similarity or phi association coefficient is computed
for binary variables a and b, such as a single pair of operational units (OUs).
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
