% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo__getMRCA.R
\name{getMRCA}
\alias{getMRCA}
\title{Getting the date of the most-recent common ancestor (MRCA) from a Bactdating tree}
\usage{
getMRCA(tr_dated)
}
\arguments{
\item{tr_dated}{A dated tree generated by function bactdate of package Bactdating (https://github.com/xavierdidelot/BactDating).}
}
\value{
The MRCA's date in a named vector (lower and upper bounds and the mean date).
}
\description{
Extracting the MRCA date from function bactdate's output.
}
