% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{is_within}
\alias{is_within}
\title{Check if a date is within an interval}
\usage{
is_within(dates, interval)
}
\arguments{
\item{dates}{A vector of dates.}

\item{interval}{An \code{Interval} object representing the date range, or something that can be coerced to an \code{Interval} object using \code{\link[=as_interval]{as_interval()}}.}
}
\value{
A logical vector indicating whether each date is within the interval.
}
\description{
This is a wrapper for the \verb{\%within\%} operator from the \code{lubridate} package.
You can create an \code{Interval} object using \code{\link[=as_interval]{as_interval()}}.
}
