% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual__labelGenerator.R
\name{labelGenerator}
\alias{labelGenerator}
\title{Creating labels for an axis at a given interval so that ticks can be denser than labels}
\usage{
labelGenerator(ticks, interval, labels)
}
\arguments{
\item{ticks}{The ticks on an axis}

\item{interval}{N, a label is laid every N ticks}

\item{labels}{A vector of labels to be placed on the axis}
}
\description{
For example, you may want the axis to be labelled by 0, 1k, 2k, etc.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
