% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__makeNamedList.R
\name{makeNamedList}
\alias{makeNamedList}
\title{Returns a one- or two-dimensional list with names for each dimension.}
\usage{
makeNamedList(names.1 = NA, names.2 = NA)
}
\arguments{
\item{names.1}{The first vector of characters as names for each level of the list}

\item{names.2:}{The second vector of characters as names for each level of the list.
e.g., an element in such a list can be assessed using x[["L1"]][["L2"]].}
}
\description{
Returns a one- or two-dimensional list with names for each dimension.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
