% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__maskMatrix.R
\name{maskMatrix}
\alias{maskMatrix}
\title{Generating a matrix consisting of 0/1 for filtering values in a target matrix.}
\usage{
maskMatrix(m, df, default = 0, keep.diag = TRUE)
}
\description{
For example, m <- m * maskMatrix(m, items_to_keep.dataframe, keep.diag = True) will replace unwanted cells with zero.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
