% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__matrix2df.R
\name{matrix2df}
\alias{matrix2df}
\title{Coverting the upper triangle of a symmetric numeric matrix into a data frame of three variables.}
\usage{
matrix2df(m, diag = FALSE, named = FALSE, small = TRUE,
  val.class = "numeric")
}
\description{
For speed, this function does not check symetry of the input matrix. It is a user's responsibility
to guarantee the input is a symmetric matrix. In addition, row names and column names must be
the same to get a correct output.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
