% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCutoff.R
\name{matrixCutoff}
\alias{matrixCutoff}
\title{Resets a value in an input matrix if it is smaller/greater than the threshold.}
\usage{
matrixCutoff(m, cutoff, comp = ">", new.val = 0)
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
