% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi__maxDaysBetweenDateRanges.R
\name{maxDaysBetweenDateRanges}
\alias{maxDaysBetweenDateRanges}
\title{Calculating maximum days between collection dates of isolates}
\usage{
maxDaysBetweenDateRanges(dates)
}
\arguments{
\item{dates}{A data frame containing three columns: Isolate, Collection_date_precision, Collection_date_min, Collection_date_max.}
}
\value{
A matrix of days between collection dates of each pair of isolates
}
\description{
This function works on the output of function estimateCollectionDates.
}
