% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxRange.R
\name{maxRange}
\alias{maxRange}
\title{Obtain the maximum range comprising two numeric vectors v1 and v2}
\usage{
maxRange(v1, v2)
}
\description{
Obtain the maximum range comprising two numeric vectors v1 and v2
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
