% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.R
\name{min}
\alias{min}
\title{Wrapper for min to handle NAs cleanly}
\usage{
min(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments (see \link[base:Extremes]{base::min}).}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The minimum value from the provided values, or NA if all values are NA.
}
\description{
\code{min} provides a wrapper for the built-in function \link[base:Extremes]{base::min}. This handles the edge case where
all values provided to min are NA, which would otherwise throw a warning and return Inf.
This is consistent with the behavior of \link[base:mean]{base::mean}.
}
\examples{
data.frame(
  x = c(1, 2, 3, NA, NA, NA),
  y = c("a", "a", "a", "b", "b", "b")
) |>
  dplyr::group_by(y) |>
  dplyr::summarise(min_x = min(x, na.rm = TRUE))

}
