% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc__mkContingencyTable.R
\name{mkContingencyTable}
\alias{mkContingencyTable}
\title{Making a contingency table from four entries}
\usage{
mkContingencyTable(a, b, c, d)
}
\description{
This programme stimulates a two-by-two contingency table.
Structure of the contingency table
      B
      1  0
A  1  a  b
   0  c  d
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
