% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinf__normaliseContigDepths.R
\name{normaliseContigDepths}
\alias{normaliseContigDepths}
\title{Median-based normalisation of contig depths}
\usage{
normaliseContigDepths(tsv)
}
\arguments{
\item{tsv}{A Tab-delimited output (a TSV file) of four columns: isolate name, contig name, contig length,
and contig depth. This file can be generated from Shovill assemblies using script extract_contig_stats.sh in
package Assembly_toolkit (https://github.com/wanyuac/Assembly_toolkit).}
}
\value{
A data frame with median-normalised depths
}
\description{
This function normalises fold coverages of contigs in a genome assembly. Depths that are close to the
median depth will be close to 1 after the normalisation.
}
\author{
Yu Wan <wanyuac@gmail.com>
}
