% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_error.R
\name{on_error}
\alias{on_error}
\title{Easy error handler}
\usage{
on_error(..., .return = NULL, .message = FALSE, .warn = FALSE)
}
\arguments{
\item{...}{A code block (typically wrapped in \code{{}}) to run and capture errors (if any).}

\item{.return}{(Optional) What is to be returned if an error occurs (instead of throwing an error).
Default is \code{NULL}}

\item{.message}{(Optional) A single logical (TRUE/FALSE) value indicating if the error message should be displayed as a message instead.
Default is \code{FALSE}}

\item{.warn}{(Optional) A single logical (TRUE/FALSE) value indicating if the error message should be displayed as a warning instead.
Default is \code{FALSE}}
}
\value{
the output of \code{...} unless an error occurs, then \code{return} instead.
}
\description{
\code{on_error} provides a simple way to handle errors by specifying a value to be returned on error and/or if a message/warning should be displayed instead.
This is especially useful when attempting to load in multiple data files where it is possible for files not to exist.
}
\examples{
on_error(stop("test"), .return = -1, .message = TRUE)
on_error(read.csv("not_A_fil3.123"), .return = NULL)
}
\seealso{
Other Utilities: 
\code{\link{for_each}()},
\code{\link{get_timezone}()},
\code{\link{on_warning}()},
\code{\link{summarise_logs}()},
\code{\link{swap}()}
}
\concept{Utilities}
