% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_warning.R
\name{on_warning}
\alias{on_warning}
\title{Easy warning handler}
\usage{
on_warning(..., .return = NULL, .message = FALSE, .stop = FALSE)
}
\arguments{
\item{...}{A code block (typically wrapped in \code{{}}) to run and capture warnings (if any).}

\item{.return}{(Optional) What is to be returned if a warning occurs (instead of throwing a warning).
Default is \code{NULL}}

\item{.message}{(Optional) A single logical (TRUE/FALSE) value indicating if the error message should be displayed as a message instead.
Default is \code{FALSE}}

\item{.stop}{(Optional) A single logical (TRUE/FALSE) value indicating if the warning message should be displayed as an error instead.
Default is \code{FALSE}}
}
\value{
the output of \code{...} unless a warning occurs, then \code{return} instead.
}
\description{
\code{on_warning} provides a simple way to handle warnings by specifying a value to be returned on warning and/or if a message/error should be displayed instead.
}
\examples{
on_warning(warning("test"), .return = -1, .message = TRUE)
on_warning(base::max(NA, na.rm = TRUE), .return = NULL)
}
\seealso{
Other Utilities: 
\code{\link{for_each}()},
\code{\link{get_timezone}()},
\code{\link{on_error}()},
\code{\link{summarise_logs}()},
\code{\link{swap}()}
}
\concept{Utilities}
