% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc__phiCoe.R
\name{phiCoe}
\alias{phiCoe}
\title{Calculating a phi association coefficient from two binary variables.}
\usage{
phiCoe(x, y = NULL)
}
\description{
Calculating a phi association coefficient from two binary variables.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
