% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo__plotCI.R
\name{plotCI}
\alias{plotCI}
\title{Plotting 95-percent confidence interval and dated tree in the result of function bactdate from the R package Bactdating.}
\usage{
plotCI(x, years, ...)
}
\arguments{
\item{x}{A dated tree generated by function bactdate of package Bactdating (github.com/xavierdidelot/BactDating).}

\item{years}{A sequence of years (integers) for ticks on the X axis. This vector can be created using the seq function.}
}
\value{
An object can be plotted in R.
}
\description{
Creating a tree figure with 95-percent confidence intervals of internal nodes' estimated dates. This function was modified
from function plot.resBactDating in github.com/xavierdidelot/BactDating/blob/master/R/methods.R.
}
