% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinf__plotGeneAccumulationCurves.R
\name{plotGeneAccumulationCurves}
\alias{plotGeneAccumulationCurves}
\title{Drawing gene-accumulation curves for a pangenome}
\usage{
plotGeneAccumulationCurves(
  L,
  point_size = 0.5,
  line_width = 0.5,
  title = "Gene accumulation",
  axis_title_size = 10,
  axis_text_size = 10
)
}
\arguments{
\item{L}{Result list of function countPanGenes}

\item{point_size}{Diameter of dots in the plot}

\item{line_width}{Width of curves}

\item{title}{Title of the plot (Default: "Gene accumulation")}

\item{axis_title_size}{Size of axis titles}

\item{axis_text_size}{Text size for axis ticks}
}
\value{
A ggplot2 object for plotting
}
\description{
This function is used for visualising results from function countPanGenes
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
