% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual__plotMetaTree.R
\name{plotMetaTree}
\alias{plotMetaTree}
\title{Drawing a metatree (i.e. tree + data + metadata) to either the screen or a file.}
\usage{
plotMetaTree(
  tree = NULL,
  heatmapData = NULL,
  barData = NULL,
  infoFile = NULL,
  blockFile = NULL,
  snpFile = NULL,
  encodedSNPMatrix = NULL,
  snp.pos = NULL,
  snp.pch = "|",
  bar.cex = 0.25,
  ladderise = NULL,
  ancestral.reconstruction = FALSE,
  heatmap.colours = rev(gray(seq(from = 0, to = 1, by = 0.1))),
  heatmapBreaks = NULL,
  heatmapDecimalPlaces = 1,
  heatmap.blocks = NULL,
  cluster = FALSE,
  cluster.method = "ward.D",
  dist.method = "euclidean",
  treeWidth = 20,
  infoWidth = 10,
  dataWidth = 50,
  barDataWidth = 10,
  blockPlotWidth = 10,
  edgeWidth = 1,
  labelHeight = 10,
  mainHeight = 100,
  barDataCol = 2,
  show.tipLabels = FALSE,
  tipLabelSize = 1,
  offset = 0,
  tip.colour.cex = 0.5,
  tipColours = NULL,
  colourNodesBy = NULL,
  infoCols = NULL,
  blwd = 5,
  block.colour = "black",
  snp.colour = "red",
  gapChar = "?",
  genome.size = 5e+06,
  genome.offset = 0,
  img.fmt = "png",
  img.name = "img",
  img.width = 640,
  img.height = 480,
  img.units = "px",
  img.res = 72,
  legend = TRUE,
  legend.pos = "bottomleft",
  pie.cex = 0.5,
  axis = FALSE,
  axisPos = 3,
  edge.color = "black",
  infoCex = 0.8,
  colLabelCex = 0.8,
  vlines.heatmap = NULL,
  vlines.heatmap.col = "gray75",
  vlines.heatmap.lty = 1,
  hlines.heatmap = NULL,
  hlines.heatmap.col = "gray75",
  hlines.heatmap.lty = 1
)
}
\description{
This script is originally developed by Dr Kathryn Holt in the University of Melbourne.
The original script (plotTree.R) was acquired from https://github.com/katholt/plotTree/tree/master on 15 Sept 2015.
There are two important changes in my version:
(1) The function name was changed from plotTree to plotMetaTree to avoid the conflict with the plotTree function in the phytools package.
(2) Options "w" and "h" were renamed to "img.width" and "img.height" to avoid the problem:
    Error in switch(units, in = res, cm = res/2.54, mm = res/25.4, px = 1) *  :
        non-numeric argument to binary operator
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
