% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo__rmSubclades.R
\name{rmSubclades}
\alias{rmSubclades}
\title{Removing IDs of sub-clades from a vector of clade names (internal nodes) of a bifurcating tree.}
\usage{
rmSubclades(tr, nodes)
}
\arguments{
\item{tr}{An input rooted bifurcating phylogenetic tree}

\item{nodes}{A character vector of internal node numbers or labels. Note that the characters should
be converted to integers if node numbers are used as input: as.integer(rmSubclades(...)).}
}
\value{
A vector of clades that do not overlap
}
\description{
This function removes subclades from a list of clades and keeps non-overlapping clades.
This function is useful for processing the output of function distsPerClade of this package.
}
