% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{rolling}
\alias{rolling}
\title{Apply a function over a rolling window}
\usage{
rolling(
  x,
  FUN = mean,
  ...,
  .width = 3,
  .direction = "backward",
  .fill = NA,
  .min_non_na = 0
)
}
\arguments{
\item{x}{A vector.}

\item{FUN}{A function to be applied to each window of \code{x}.
Default is \code{mean}.}

\item{...}{Additional arguments to be passed to \code{FUN}.}

\item{.width}{A numeric value indicating the width of the window.
See \code{\link[zoo:rollapply]{zoo::rollapply()}} for more details.
Default is \code{3}.}

\item{.direction}{A character string or numeric value indicating the direction of the window.
Options are \code{"backward"} (1), \code{"forward"} (-1), or \code{"center"} (0).
See the \code{align} argument in \code{\link[zoo:rollapply]{zoo::rollapply()}} for more details.
Default is \code{"backward"}.}

\item{.fill}{(Optional) A single value to be used for filling in any values that are outside of the window.
See \code{\link[zoo:rollapply]{zoo::rollapply()}} for more details.
Default is \code{NA}.}

\item{.min_non_na}{(Optional) A single numeric value indicating the minimum number of observations required to
compute a value.
Default is \code{0}.}
}
\value{
A vector of the same length as \code{x} and type of the output of \code{FUN}.
}
\description{
\code{rolling} applies a funtion (\code{FUN}) over a moving window (see \code{.width} and \code{.direction}) for each element of \code{x} using \code{\link[zoo:rollapply]{zoo::rollapply()}}.
}
\details{
\itemize{
\item Be sure to fill any gaps in \code{x} and ensure \code{x} is sorted, otherwise the window may be applied incorrectly.
\item Values outside of the window are replaced with \code{.fill}.
\item \code{.min_non_na} is passed to \code{\link[=do_if_enough]{do_if_enough()}} to only apply \code{FUN} if enough values are present in the window.
}
}
\examples{
x <- c(1, 2, 3, 4, 5)
x |> rolling(mean, .width = 2)
x |> rolling(mean, .width = 2, .direction = "forward", .fill = -1)
}
