% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figure.R
\name{save_figure}
\alias{save_figure}
\title{Save a ggplot2 figure for publishing}
\usage{
save_figure(
  gg,
  out_path,
  page_width = 6.5,
  base_height = 5,
  taller = 0,
  units = "in",
  quality = "high",
  ...
)
}
\arguments{
\item{gg}{a single ggplot object you would like to save.}

\item{out_path}{A single character value indicating where to save the plot to (i.e. "./plots/my_figure_name.png")}

\item{page_width}{A single numeric value indicating the width of the page (minus margins) that the image is intended for.
Default is 6.5".}

\item{base_height}{A single numeric value indicating the height of the figure (+\code{taller}).
Default is 5".}

\item{taller}{A single numeric value indicating how much to add/remove to the plot \code{base_height}.
Default is 0.}

\item{units}{A single character value equal to "in" (default) or "cm" indicating the units of \code{page_width}, \code{base_height}, and \code{taller}.}

\item{quality}{A single character value equal to "high" (dpi = 300), "medium" (dpi = 200), or "low" (dpi = 100) indicating the output quality of the figure.
OR a single numeric value equal to the desired dpi.
Text sizes may need to be adjusted in relavant geoms/themes for different quality levels.}

\item{...}{(Optional) addition arguments passed on to ggplot2::ggsave()}
}
\value{
invisibly returns \code{gg}
}
\description{
\code{save_figure()} is a wrapper around \code{ggplot2::ggsave()} that saves a high quality 5x6.5 figure by default.
You can adjust the sizing from there by changing \code{taller} and \code{page_width}.
}
\examples{
gg <- ggplot2::ggplot() +
  ggplot2::geom_line(
    data = data.frame(x = 1:10, y = (0:9)^2),
    ggplot2::aes(x, y)
  )
# save_figure(gg, "./test.png", taller = 1)
}
\concept{Data Visualisation}
