% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{seq.Interval}
\alias{seq.Interval}
\title{Extend seq() to work with Intervals}
\usage{
\method{seq}{Interval}(interval, by = "auto", ...)
}
\arguments{
\item{interval}{A \code{Interval} object representing the date range.
Multiple intervals will be looped over using \code{sapply()} with the same \code{by} and \code{...} arguments.}

\item{by}{A string indicating the interval to use.
Can be one of "auto", "1 seconds", "1 minutes", "1 hours", or "1 days".
If "auto", the interval will be automatically determined based length of the interval.
Default is "auto".}

\item{...}{Additional arguments passed to \code{seq()}.}
}
\value{
A sequence of dates within the interval.
}
\description{
\code{Interval} objects can be created using \verb{\%--\%} from the \code{lubridate} package, or the handy wrapper \code{\link[=as_interval]{as_interval()}}.
These objects represent a range of time, so it makes sense to be able to generate a sequence of dates within that range.

\code{\link[=seq.Interval]{seq.Interval()}} extends \code{\link[=seq]{seq()}} to work with \code{Interval} objects.
}
