% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDepth.R
\name{seqDepth}
\alias{seqDepth}
\title{Calculating sequencing depths from the output of 'seqkit stats'}
\usage{
seqDepth(reads, len, srt = TRUE)
}
\arguments{
\item{reads}{A data frame imported from the output file of 'seqkit stats'. The first column of
input file names must not contain paths. For example, sample_1.fastq.gz is valid whereas
reads/filtered/sample_1.fastq.gz is invalid. Each FASTQ file should follow the pattern *_{1, 2}.fastq.gz.}

\item{len}{Length (bp) of the reference genome}

\item{srt}{Whether to sort rows by sequencing depths (default: TRUE)}
}
\value{
A data frame of isolate names and sequencing dapths
}
\description{
This function calculates the sequencing depth of each genome using the output of
command 'seqkit stats' (https://github.com/shenwei356/seqkit)
}
