% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_as_df.R
\name{sf_as_df}
\alias{sf_as_df}
\title{Convert an sf object to a data.frame}
\usage{
sf_as_df(sf_obj, keep_coords = FALSE)
}
\arguments{
\item{sf_obj}{An object of class \code{sf}, typically created by \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{keep_coords}{A logical value indicating if the coordinates should be kept when dropping the geometry}
}
\value{
A \code{data.frame} object without the geometry column.
}
\description{
This function removes the geometry from an sf object, effectively converting it back to a data.frame.
}
\examples{
cities <- data.frame(
  name = c("Nanaimo", "Port Moody", "Prince George"),
  x = c(-124.0531, -122.8519, -122.7949),
  y = c(49.1633, 49.2844, 53.8934)
)
cities_sf <- cities |>
  sf::st_as_sf(coords = c("x", "y"), crs = "WGS84")
sf_as_df(cities_sf)
sf_as_df(cities_sf, keep_coords = TRUE)
}
