% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_number.R
\name{shorten_number}
\alias{shorten_number}
\title{Shorten a number to a string with a unit prefix (e.g. "1K", "10M", "100B")}
\usage{
shorten_number(
  x,
  decimals = 2,
  reverse = FALSE,
  thresholds = list(K = 1000, M = 1e+06, B = 1e+09, T = 1e+12)
)
}
\arguments{
\item{x}{A vector of numeric values to shorten, or character values to lengthen (if \code{reverse = TRUE}).}

\item{decimals}{The number of decimal places to round \code{x} to if \code{reverse = FALSE}. Default is \code{2}.}

\item{reverse}{A logical indicating if the thresholds should be used in reverse order.}

\item{thresholds}{A named list of thresholds.
The names of the list are used as the unit prefix.
The values of the list must be powers of 10.
Default is \code{list("K" = 1e3, "M" = 1e6, "B" = 1e9, "T" = 1e12)}.}
}
\value{
A shortened string representation of \code{x} if \code{reverse = FALSE},
or a numeric vector of lengthened \code{x} values if \code{reverse = TRUE}.
}
\description{
Provided numbers will be divided by their highest met threshold and rounded to the specified number of decimal places.
These values will then be converted to characters and appended with the appropriate threshold symbol.
}
\details{
The reverse can be performed by setting \code{reverse = TRUE} and providing a character vector to \code{x} (i.e. "1K" -> 1000).
}
\examples{
x <- c(
   1,
   100,
   1000,
   10000,
   100000,
   1000000,
   10000000,
   100000000,
   1000000000,
   10000000000
 )
shorten_number(x)

x <- c("1", "100", "1K", "10K", "100K", "1M", "10M", "100M", "1B", "10B")
shorten_number(x, reverse = TRUE)
}
