% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silence.R
\name{silence}
\alias{silence}
\title{Silence unwanted output}
\usage{
silence(
  this_please,
  errors = TRUE,
  warnings = TRUE,
  messages = TRUE,
  output = TRUE
)
}
\arguments{
\item{this_please}{Some code to run silently.
If running multiple lines use \code{{....}}}

\item{errors}{A logical value indicating if errors should be silenced.
Default is \code{TRUE}.}

\item{warnings}{A logical value indicating if warnings should be silenced.
Default is \code{TRUE}.}

\item{messages}{A logical value indicating if messages should be silenced.
Default is \code{TRUE}.}

\item{output}{A logical value indicating if output from \code{print()}, \code{cat()}, and \code{writeLines()} should be silenced.
Default is \code{TRUE}.}
}
\value{
The output from \code{this_please}
}
\description{
This function wraps \code{tryCatch} and suppresses some or all of the following:
\itemize{
\item errors
\item warnings
\item messages
\item output from \code{print()}
\item output from \code{cat()}
\item output from \code{writeLines()}
}
}
\examples{
silence(1/0)
silence(warning("test"))
silence(message("test"))
silence(print("test"))
silence(cat("test"))
silence(1 + "a")
}
