% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc__simCoe.R
\name{simCoe}
\alias{simCoe}
\title{Simpson's similarity coefficient between two binary vectors}
\usage{
simCoe(x, y)
}
\arguments{
\item{x}{The first binary vector}

\item{y}{The second binary vector}
}
\description{
Simpson's similarity coefficient between two binary vectors
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
