% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_date_range.R
\name{split_date_range}
\alias{split_date_range}
\title{Split a date range into multiple, shorter date ranges}
\usage{
split_date_range(date_range, max_duration = "90 days", as_list = FALSE)
}
\arguments{
\item{date_range}{A length-2 vector of Dates or POSIXcts representing the start and end of a date range.}

\item{max_duration}{A string representing the maximum duration of each split date range. Defaults to "90 days".}

\item{as_list}{A logical indicating if the output should be a list of data frames (one for each split date range) or a single data frame with all split date ranges. Defaults to \code{FALSE}.}
}
\value{
If \code{as_list = FALSE} (default), a data frame with columns \code{start} and \code{end} representing the start and end of each split date range.
If \code{as_list = TRUE}, a list of data frames, each with columns \code{start} and \code{end} representing the start and end of each split date range.
}
\description{
Split a date range into multiple, shorter date ranges
}
