% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__splitVec.R
\name{splitVec}
\alias{splitVec}
\title{Splitting a vector into a list of n vectors}
\usage{
splitVec(v, n)
}
\arguments{
\item{v}{Input vector to be broken}

\item{n}{Number of vectors to be generated from v}
}
\value{
An unnamed list of n vectors
}
\description{
This function splits a vector into a list of n vectors
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
