% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{Swap out values in a vector}
\usage{
swap(x, what, with)
}
\arguments{
\item{x}{Vector of values to be have certain values swapped out.}

\item{what}{One or more values to be replaced with \code{with} throughout \code{x}.}

\item{with}{A single value to replace \code{what} for throughout \code{x}.}
}
\value{
a vector of \code{x} where all instances of \code{what} are replaced with \code{with}
}
\description{
\code{swap} provides a simple way to switch out certain values in a vector. It is useful for replacing NA's, Infinites, and erroneous values.
}
\examples{
swap(c(-20:20, NA), what = NA, with = -1)
swap(c(-20:20, Inf), what = Inf, with = NA)
swap(c(-20:20), what = Inf, with = NA)
}
\seealso{
Other Utilities: 
\code{\link{for_each}()},
\code{\link{on_error}()}
}
\concept{Utilities}
