% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__tab2df.R
\name{tab2df}
\alias{tab2df}
\title{Table to data frame}
\usage{
tab2df(t, new_names = NA)
}
\arguments{
\item{t}{Input table}

\item{new_names}{A character vector of two elements for new column names. Default: Name, Count.}
}
\description{
Converts a table from function "table" to a two column data frame.
}
