% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc__testIndependency.R
\name{testIndependency}
\alias{testIndependency}
\title{Testing the independency of two variables according to a contingency table}
\usage{
testIndependency(m, method = "chisq", alt = "two.sided")
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
