% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__triangleToSquareMatrix.R
\name{triangleToSquareMatrix}
\alias{triangleToSquareMatrix}
\title{Converting a data frame representing a triangle matrix into a square matrix.}
\usage{
triangleToSquareMatrix(d, diag_val = NA, sort_id = FALSE,
  sort_decreasing = FALSE)
}
\arguments{
\item{d}{Input data frame}

\item{diag_val}{The value on the diagonal (Default: NA)}

\item{sort_id}{A logical flag specifying whether to sort row and column names of the output matrix}

\item{sort_decreasing}{A logical flag specifying whether to sort row and column names in a decreasing order}
}
\value{
A synmetric matrix
}
\description{
This function retrieves IDs from the first two columns and values from the third column.
Note that this function assumes the diagonal of the output matrix is comprised of a constant.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
