% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncate}
\alias{truncate}
\title{Truncate a numeric vector to a certain number of decimal places}
\usage{
truncate(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector to be truncated}

\item{digits}{A single numeric value indicating the number of decimal places to round to
Default is \code{0}}
}
\value{
A numeric vector of \code{x} truncated to \code{digits} decimal places
}
\description{
Truncate a numeric vector to a certain number of decimal places
}
