% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc__upperTriangleToSquareMatrix.R
\name{upperTriangleToSquareMatrix}
\alias{upperTriangleToSquareMatrix}
\title{Converting a data frame representing an upper triangle matrix (without diagonal entries)
into the original square matrix.}
\usage{
upperTriangleToSquareMatrix(d, diag_val = 0)
}
\arguments{
\item{d}{Input data frame}

\item{diag_val}{The value on the diagonal (Default: NA)}
}
\value{
A synmetric matrix
}
\description{
This function retrieves IDs from the first two columns and values from the third
column. Note that this function assumes the diagonal of the output matrix is comprised of a
constant.creases.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
