% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_database.R
\name{write_to_database}
\alias{write_to_database}
\title{Write data to a database}
\usage{
write_to_database(
  db,
  table_name,
  new_data,
  primary_keys,
  unique_indexes = NULL,
  insert_new = TRUE,
  update_duplicates = FALSE
)
}
\arguments{
\item{db}{A database connection or path to database (if sqlite or duckdb file extension).}

\item{table_name}{A character string specifying the table to write to.}

\item{new_data}{A data frame to write to the table.}

\item{primary_keys}{A character vector of column names to use as the primary key, the main identifier of individual rows in a table.
Multiple columns can be specified and uniqueness will be assessed based on the combination of columns.
(e.g. \code{primary_keys = c("col1", "col2")} will add a primary key on the combination of \code{col1} and \code{col2}).}

\item{unique_indexes}{A list of character vector(s) of column names to use as unique indexes.
These will be added to the table, in addition to the primary key.
Multiple columns can be specified and uniqueness will be assessed based on the combination of columns.
(e.g. \code{unique_indexes = list(c("col1", "col2"))} will add a unique index on the combination of \code{col1} and \code{col2}.).
If \code{NULL} (the default), no unique indexes will be added.}

\item{insert_new}{A logical value indicating if new data should be inserted into the existing table.
If \code{FALSE}, no new data will be inserted, only existing rows will be updated if \code{update_duplicates = TRUE}.}

\item{update_duplicates}{A logical value indicating if existing rows should be updated with new data.
If \code{FALSE}, no existing rows will be updated. Only new data will be inserted if \code{insert_new = TRUE}.}
}
\value{
An invisible db connection.
}
\description{
Write data to a database, with options to insert new data, update existing data, and create a new table.
}
\details{
It will create the table and insert \code{new_data} if the table does not already exist.
Otherwise:
\itemize{
\item If \code{insert_new = TRUE}, \code{new_data} will be inserted into the existing table.
\item If \code{update_duplicates = TRUE}, existing data will be updated with the overlap of \code{new_data}.
}
}
