% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo__writePHYLIPmatrix.R
\name{writePHYLIPmatrix}
\alias{writePHYLIPmatrix}
\title{Create a PHYLIP-formatted distance matrix from the output of function makeFastANIDistMatrix}
\usage{
writePHYLIPmatrix(M, f)
}
\arguments{
\item{M}{The distance matrix with row and column names}

\item{f}{Path of the output file}
}
\description{
This function takes as input a square matrix and write a PHYLIP-formatted matrix file.
See http://evolution.genetics.washington.edu/phylip/doc/distance.html for an example of the format.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
