% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{as.data.frame.Interval}
\alias{as.data.frame.Interval}
\title{Convert an Interval object to a data frame}
\usage{
\method{as.data.frame}{Interval}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An \code{Interval} object.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{...}{Additional arguments passed to \code{data.frame()}.}
}
\value{
A data frame with two columns: \code{start} and \code{end}.
}
\description{
Split an \code{Interval} object into a data frame with two columns: \code{start} and \code{end}.
}
