% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamp.R
\name{clamp}
\alias{clamp}
\title{Clamp a value to a certain range}
\usage{
clamp(x, range)
}
\arguments{
\item{x}{A numeric vector of value(s)}

\item{range}{A numeric vector of length two to clamp \code{x} to}
}
\value{
A numeric vector of \code{x} with values less than \code{min(range)} set to \code{min(range)} and values greater than \code{max(range)} set to \code{max(range)}
}
\description{
Clamp a value to a certain range
}
