% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between common units}
\usage{
convert_units(x, from, to, keep_units = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{from, to}{A string indicating the current/desired unit of \code{x}.
Run \code{\link[units:valid_udunits]{units::valid_udunits()}} for a list of valid units for your system.}

\item{keep_units}{A logical value indicating whether to keep the units after conversion (default is \code{FALSE}).}

\item{...}{Additional arguments passed on to \code{\link[units:units]{units::set_units()}}}
}
\value{
A numeric vector with the same length/type as \code{x}.
If \code{keep_units = TRUE}, the units attribute is preserved
}
\description{
\code{convert_units} is a wrapper around \code{\link[units:units]{units::set_units()}} to switch between common measurement units.
}
\examples{
convert_units(c(1, 2, 3), from = "cm", to = "km")
convert_units(c(1, 2, 3), from = "degF", to = "degC")
convert_units(c(1, 2, 3), from = "ft", to = "m")
convert_units(c(1, 2, 3), from = "oz", to = "L")
convert_units(c(1, 2, 3), from = "m/s", to = "km/h")
convert_units(c(1, 2, 3), from = "ug/m3", to = "mg/km3")
}
