% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if.R
\name{do_if}
\alias{do_if}
\title{Apply a function if a condition is met}
\usage{
do_if(x, .do, .if = TRUE, ..., .return = NA)
}
\arguments{
\item{x}{Something that \code{.do} can be applied to.}

\item{.do}{A function to be applied to \code{x}.}

\item{.if}{A logical value indicating if \code{.do} should be applied or not.}

\item{...}{Additional arguments to be passed to \code{.do}.}

\item{.return}{What to return instead of the output of \code{.do} if \code{.if == FALSE}.}
}
\value{
The output of \code{.do} applied to \code{x} if \code{.if == TRUE}, \code{.return} otherwise
}
\description{
Apply a function if a condition is met
}
\examples{
do_if(c(1, 2, 3), mean, .if = TRUE)
do_if(c(1, 2, 3), mean, .if = FALSE)
do_if(c(1, 2, 3), mean, .if = FALSE, .return = -1)
}
