% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi__estimateCollectionDates.R
\name{estimateCollectionDates}
\alias{estimateCollectionDates}
\title{Estimating dates of sample collection}
\usage{
estimateCollectionDates(dates)
}
\arguments{
\item{dates}{A data frame of four columns: Isolate, Collection_date (YYYY-MM-DD, YYYY, YYYY-MM), Collection_date_precision (Y, YM, YMD), Receipt_date (YYYY-MM-DD).
Unknown dates must be NA.}
}
\value{
A data frame of the minimum and maximum collection dates of each isolate.
}
\description{
This function estimates the range of collection dates "YYYY-MM" for each bacterial isolate based on a four-column data frame.
}
