% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sf_coords.R
\name{extract_sf_coords}
\alias{extract_sf_coords}
\title{Extract coordinates from an sf object and add them as columns}
\usage{
extract_sf_coords(sf_obj, keep_sf = TRUE)
}
\arguments{
\item{sf_obj}{An object of class \code{sf}, typically created by \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{keep_sf}{A logical value indicating if the return object should remain an \code{sf} object or be converted to a \code{data.frame}}
}
\value{
An \code{sf} object with two additional columns: \code{lng} and \code{lat} for longitude and latitude respectively.
}
\description{
Extract coordinates from an sf object and add them as columns
}
\examples{

cities <- data.frame(
  name = c("New York", "Los Angeles", "Chicago"),
  lat = c(40.71278, 34.05224, 41.87811),
  lon = c(-74.0060, -118.2437, -87.6298)
)
cities_sf <- cities |> 
  sf::st_as_sf(coords = c("lon", "lat"), crs = "WGS84")
cities_sf |>
  extract_sf_coords()
}
