% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sf_coords.R
\name{extract_sf_coords}
\alias{extract_sf_coords}
\title{Extract coordinates from an sf object and add them as columns}
\usage{
extract_sf_coords(sf_obj, keep_sf = TRUE)
}
\arguments{
\item{sf_obj}{An object of class \code{sf}, typically created by \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{keep_sf}{A logical value indicating if the return object should remain an \code{sf} object or be converted to a \code{data.frame}}
}
\value{
An \code{sf} object with two additional columns: \code{lng} and \code{lat} for longitude and latitude respectively.
}
\description{
Extract coordinates from an sf object and add them as columns
}
\examples{

cities <- data.frame(
  name = c("Nanaimo", "Port Moody", "Prince George"),
  x = c(-124.0531, -122.8519, -122.7949),
  y = c(49.1633, 49.2844, 53.8934)
)
cities_sf <- cities |>
  sf::st_as_sf(coords = c("x", "y"), crs = "WGS84")
cities_sf |>
  extract_sf_coords()
}
