% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timezone.R
\name{get_timezone}
\alias{get_timezone}
\title{Wrapper for looking up timezone of locations from lat/lng coords using the \code{lutz} package}
\usage{
get_timezone(lng, lat, method = "accurate", ...)
}
\arguments{
\item{lng}{Vector of numeric values indicating location longitudes (decimal degrees) to lookup.}

\item{lat}{Vector of numeric values indicating location latitudes (decimal degrees) to lookup.}

\item{method}{A single character value indicating the lookup method to use (either "accurate" or "fast"). See \link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords} for more details.}

\item{...}{(Optional) Additional paramaters to pass to \link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords}}
}
\value{
a character vector with the same length as \code{lat} and \code{lng} indicating the locations likely timezone.
}
\description{
\code{get_timezone} uses the \code{lutz} package to lookup the timezone of locations from lat/lng coords. See \link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords} for more details.
}
\examples{
get_timezone(-105.053144, 69.116178, method = "fast")
}
\seealso{
Other Utilities: 
\code{\link{for_each}()},
\code{\link{on_error}()},
\code{\link{summarise_logs}()},
\code{\link{swap}()}
}
\concept{Utilities}
