% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_step}
\alias{log_step}
\title{Log a message with optional timestamp and formatting}
\usage{
log_step(
  message,
  header = FALSE,
  time = !header,
  time_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = NULL,
  quiet = FALSE
)
}
\arguments{
\item{message}{A character value of the message to be logged}

\item{header}{A logical value indicating if the message should be formatted as a header}

\item{time}{A logical value indicating if the current time should be prepended to the log message}

\item{time_format}{A character value indicating the format of the timestamp}

\item{tz}{A character value indicating the time zone to use for the timestamp}

\item{quiet}{A logical value indicating if the message should be printed}
}
\value{
a character vector of the formatted message
}
\description{
Log a message with optional timestamp and formatting
}
\examples{
log_step("My Awesome Script", time = FALSE, header = TRUE)
log_step("Step 1...")
# Do something
log_step("Step 2...")
# Do something else
log_step("Complete")
}
