% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{max}
\alias{max}
\title{Wrapper for max to handle NAs cleanly}
\usage{
max(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments (see \link[base:Extremes]{base::max}).}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The maximum value from the provided values, or NA if all values are NA.
}
\description{
\code{max} provides a wrapper for the built-in function \link[base:Extremes]{base::max}. This handles the edge case where
all values provided to max are NA, which would otherwise throw an warning and return Inf.
This is consitent with the behavior of \link[base:mean]{base::mean}.
}
\examples{
data.frame(
  x = c(1, 2, 3, NA, NA, NA),
  y = c("a", "a", "a", "b", "b", "b")
) |>
  dplyr::group_by(y) |>
  dplyr::summarise(max_x = max(x, na.rm = TRUE))

}
