% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min.R
\name{max}
\alias{max}
\alias{min}
\title{Get the maximum/minimum value while handling NAs cleanly}
\usage{
max(..., na.rm = FALSE)

min(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character vectors (see \link[base:Extremes]{base::max}).}

\item{na.rm}{A logical value indicating whether NA values should be removed.}
}
\value{
A single value with the same type as \code{...}
}
\description{
\code{max}/\code{min} provides a wrapper for the built-in functions \link[base:Extremes]{base::max}/\link[base:Extremes]{base::min} that handles NAs the same as \link[base:mean]{base::mean}.
}
\details{
This handles the case where all values provided to max/min are NA, which normally returns Inf/-Inf and throws a warning (see examples).
}
\examples{
# Edge case where all values are NA
base::max(c(NA, NA, NA)) # Returns NA
base::max(c(NA, NA, NA), na.rm = TRUE) # Returns -Inf and throws warning
base::min(c(NA, NA, NA), na.rm = TRUE) # Returns Inf and throws warning
max(c(NA, NA, NA)) # Returns NA
max(c(NA, NA, NA), na.rm = TRUE) # Returns NA
min(c(NA, NA, NA), na.rm = TRUE) # Returns NA

# Example usage with typical dplyr pipeline
data.frame(
  x = c(1, 2, 3, NA, NA, NA),
  y = c("a", "a", "a", "b", "b", "b")
) |>
  dplyr::group_by(y) |>
  dplyr::summarise(
    base_max_x = base::max(x, na.rm = TRUE), # for comparison
    base_min_x = base::min(x, na.rm = TRUE), # for comparison
    max_x = max(x, na.rm = TRUE),
    min_x = min(x, na.rm = TRUE)
  )

}
