% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Calculate the stitistical mode (most common value)}
\usage{
mode(..., na.rm = TRUE)
}
\arguments{
\item{...}{numeric or character arguments (see \link[base:Extremes]{base::max}).}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The most common value from the provided values, or NA if all values are NA.
}
\description{
\code{mode} determines the most common value in a vector while cleanly handling NAs
}
\examples{
data.frame(
  x = c(1, 1, 3, NA, NA, NA),
  y = c("a", "a", "a", "b", "b", "b")
) |>
  dplyr::group_by(y) |>
  dplyr::summarise(mode_x = mode(x, na.rm = TRUE))

}
