% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{rolling}
\alias{rolling}
\title{Perform a rolling calculation on a vector if enough values are provided.}
\usage{
rolling(
  x,
  FUN = mean,
  ...,
  .width = 3,
  .direction = "backward",
  .fill = NA,
  .min_non_na = 0
)
}
\arguments{
\item{x}{A vector of values}

\item{FUN}{(Optional) A function to be applied to each window of \code{x}.
The default is \code{mean}.}

\item{...}{(Optional) Additional arguments to be passed to \code{FUN}}

\item{.width}{(Optional) A single numeric value indicating the width of the window.
The default is \code{3}.}

\item{.direction}{(Optional) A single character string or numeric value indicating the direction of the window.
The default is \code{"backward"}.
Options are \code{"backward"} (1), \code{"forward"} (-1), or \code{"center"} (0).}

\item{.fill}{(Optional) A single value to be used for filling in any values that are outside of the window.
The default is \code{NA}.}

\item{.min_non_na}{(Optional) A single numeric value indicating the minimum number of observations required to
compute a value.
The default is \code{0}.}
}
\description{
Perform a rolling calculation on a vector if enough values are provided.
}
\examples{
x <- c(1, 2, 3, 4, 5)
x |> rolling(mean, .width = 2)
x |> rolling(mean, .width = 2, .direction = "forward", .fill = -1)
}
