% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figure.R
\name{save_figure}
\alias{save_figure}
\title{Save a ggplot2 figure}
\usage{
save_figure(gg, out_path, taller = 0, page_width = 6.5, quality = "high", ...)
}
\arguments{
\item{gg}{a single ggplot object you would like to save.}

\item{out_path}{A single character value indicating where to save the plot to (i.e. "./plots/my_figure_name.png")}

\item{taller}{A single numeric value indicating how many inches to add/remove to the output plot height. Default makes a 5"x6.5" image.}

\item{page_width}{A single numeric value indicating the width of the page (minus margins) in inches that the image is intended for.}

\item{quality}{A single character value equal to "high" (dpi = 300), "medium" (dpi = 200), or "low" (dpi = 100) indicating the output quality of the figure.
OR a single numeric value equal to the desired dpi.
Text sizes may need to be adjusted in relavant geoms/themes for different quality levels.}

\item{...}{(Optional) addition arguments passed on to ggplot2::ggsave()}
}
\value{
invisibly returns \code{gg}
}
\description{
\code{save_figure()} is a wrapper around \code{ggplot2::ggsave()} that saves a high quality 5x7 figure by default.
You can adjust the sizing from there by changing \code{taller} and \code{page_width}.
}
\examples{
gg <- ggplot2::ggplot() +
  ggplot2::geom_line(
    data = data.frame(x = 1:100, y = (0:99)^2),
    ggplot2::aes(x, y)
  )
# save_figure(gg, "./test.png", taller = 1)
}
\concept{Data Visualisation}
