% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_as_df.R
\name{sf_as_df}
\alias{sf_as_df}
\title{Convert an sf object to a data.frame}
\usage{
sf_as_df(sf_obj, keep_coords = FALSE)
}
\arguments{
\item{sf_obj}{An object of class \code{sf}, typically created by \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{keep_coords}{A logical value indicating if the coordinates should be kept when dropping the geometry}
}
\value{
A \code{data.frame} object without the geometry column.
}
\description{
This function removes the geometry from an sf object, effectively converting it back to a data.frame.
}
\examples{
cities <- data.frame(
  name = c("New York", "Los Angeles", "Chicago"),
  lat = c(40.71278, 34.05224, 41.87811),
  lon = c(-74.0060, -118.2437, -87.6298)
)
cities_sf <- cities |>
  sf::st_as_sf(coords = c("lon", "lat"), crs = "WGS84")
sf_as_df(cities_sf)
sf_as_df(cities_sf, keep_coords = TRUE)
}
