% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_logs.R
\name{summarise_logs}
\alias{summarise_logs}
\title{Summarise logs stored from repeated \code{log} calls}
\usage{
summarise_logs(logs)
}
\arguments{
\item{logs}{A list of \code{log} values}
}
\value{
a character vector indicating the summary of the time taken between each log call
}
\description{
\code{summarise_logs} takes a list of \code{log} values and returns a summary of the time taken between each log call.
}
\examples{
logs <- list(
  log_step("test", quiet = TRUE),
  log_step("test", quiet = TRUE),
  log_step("test", quiet = TRUE),
  log_step("test", quiet = TRUE),
  log_step("test", quiet = TRUE),
  log_step("test", quiet = TRUE)
)
summarise_logs(logs)
}
\seealso{
Other Utilities: 
\code{\link{swap}()}
}
\concept{Utilities}
