% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_logs.R
\name{summarise_logs}
\alias{summarise_logs}
\title{Log overall and individual run time for repeated \code{\link[=log_step]{log_step()}} calls}
\usage{
summarise_logs(logs, save_to = NULL)
}
\arguments{
\item{logs}{A list of \code{\link[=log_step]{log_step()}} return values}

\item{save_to}{A character value indicating the path to a text file to save the summary of the logs to.
Default is \code{NULL} (do not save a log file).}
}
\value{
a character vector indicating the summary of the time taken between each log call
}
\description{
\code{summarise_logs} takes a list of log entries from \code{\link[=log_step]{log_step()}} and outpus a message with a summary of the time taken between each log call.
}
\examples{
logs <- log_step("Example usage of log_summarise", header = TRUE)
logs$step_one <- log_step("Step-1...")
# Do something
logs$step_two <- log_step("Step-2...")
# Do something else
logs$step_three <- log_step("Step-3...")
# Do something else
logs$done <- log_step("Complete")
# Summarise run times and write to file
summarise_logs(logs, save_to = tempfile())
}
\seealso{
Other Utilities: 
\code{\link{swap}()}
}
\concept{Utilities}
