% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinf__summariseSeqkitPEReadStats.R
\name{summariseSeqkitPEReadStats}
\alias{summariseSeqkitPEReadStats}
\title{Summarising statistics of paired-end read sets}
\usage{
summariseSeqkitPEReadStats(
  rs,
  ref_len = 5e+06,
  ext = ".fastq.gz",
  suf_R = FALSE,
  srt = "increasing"
)
}
\arguments{
\item{rs}{Tab-delimited output (a TSV file) from command 'seqkit stats'}

\item{ref_len}{Length of the reference genome in base pairs}

\item{ext}{Filename extension of FASTQ files in the input TSV file. For example, '.fastq.gz' or '.fq.gz'.}

\item{suf_R}{A logical value indicating whether 'R' is used in the filename suffices. For instance, suf_R = TRUE when read files are ended with '_R1.fastq.gz' and '_R2.fastq.gz'.}

\item{srt}{A string with values "decreasing", "increasing", or NULL (no sorting) indicating whether the output data frame will be sorted in a specific order for sequencing depths and isolate names.}
}
\value{
A data frame with summary statistics including sequencing depths.
}
\description{
This function summarises statistics from seqkit for paired-end read sets.
}
\author{
Yu Wan <wanyuac@126.com>
}
